make discussion thread view into its own component
