+++ app/components/discussion-thread-view/template.html
@@ -1,4 +1,4 @@
-<div class="container" ng-controller="DiscussionCtrl">
+<div class="container">
   <h3 class="ph-no-margin-top">
     <inline-editable
       original-content="discussion.title"
